Enum E2E test: assert on result from `storeObjectWithEnum` instead of doing a further GET via `getObjectWithEnum`
Move `ObjectWithEnumBuilder` into its own file
